requirements.txt
fastapi
uvicorn
pdfplumber
pymupdf
opencv-python-headless
pytesseract
Pillow
pydantic
python-multipart
reportlab